Add field to struct window.
Think I'll need this field to implement the duplicate discard. It
can be used to check how old the window is. I was thinking that I could
use it to determine whether to accept an older frame outside the window
without moving the window or not. If a frame arrives towards the left
of the window, if the time difference between the age of the window
given by @last_jiffies and the current jiffies is greater than
ENTRY_FORGET_TIME, then we can shift the window to contain that frame.